opensim-cmd: Moved subcommands into their own files in a subcommands/ directory
