+++ Applications/opensim-cmd/cli_helpers.h
@@ -1,0 +1,44 @@
+/* -------------------------------------------------------------------------- *
+ *                       OpenSim:  opensim-cmd.cpp                            *
+ * -------------------------------------------------------------------------- *
+ * The OpenSim API is a toolkit for musculoskeletal modeling and simulation.  *
+ * See http://opensim.stanford.edu and the NOTICE file for more information.  *
+ * OpenSim is developed at Stanford University and supported by the US        *
+ * National Institutes of Health (U54 GM072970, R24 HD065690) and by DARPA    *
+ * through the Warrior Web program.                                           *
+ *                                                                            *
+ * Copyright (c) 2005-2017 Stanford University and the Authors                *
+ * Author(s): Frank C. Anderson, Ayman Habib, Chris Dembia                    *
+ *                                                                            *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may    *
+ * not use this file except in compliance with the License. You may obtain a  *
+ * copy of the License at http://www.apache.org/licenses/LICENSE-2.0.         *
+ *                                                                            *
+ * Unless required by applicable law or agreed to in writing, software        *
+ * distributed under the License is distributed on an "AS IS" BASIS,          *
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.   *
+ * See the License for the specific language governing permissions and        *
+ * limitations under the License.                                             *
+ * -------------------------------------------------------------------------- */
+
+#ifndef _opensim_cli_helpers_h_
+#define _opensim_cli_helpers_h_
+
+namespace OpenSimCmd {
+
+bool is_arg(const char* c, const char* cmp);
+
+template<typename... T>
+bool is_arg(const char* c, const char* cmp1, T... cmp) {
+    return is_arg(c, cmp1) or is_arg(c, cmp...);
+}
+
+bool is_help_arg(const char* c);
+
+bool starts_with(const char* pre, const char* s);
+
+bool extract_opt_arg(int& argc, const char**& argv, const char* pre,  const char*& out);
+
+}
+
+#endif // _opensim_cli_helpers_h
